Update whychoose section in franchise page
