+++ pages/franchise.js
@@ -1,6 +1,6 @@
 import ContactUsBanner from "@components/ContactUsBanner";
-import TextImgGrid from "@components/TextImgGrid";
-import TextImgGridMobile from "@components/TextImgGridMobile";
+import TextImgGrid from "@components/franchise/TextImgGrid";
+import TextImgGridMobile from "@components/franchise/TextImgGridMobile";
 import FranchiseProcess from "@components/franchise/FranchiseProcess";
 import TopBanner from "@components/TopBanner";
